Halfway there

+++ src/Ingredient.java
@@ -1,8 +1,8 @@
 public class Ingredient {
     public String name;
-    public int amount;
+    public float amount;
 
-    Ingredient(String name, int amount){ //Takes the input name and amount and makes them one ingredient object
+    Ingredient(String name, float amount){ //Takes the input name and amount and makes them one ingredient object
        this.name = name;
        this.amount = amount;
     }
